progress 2

+++ lib/app/providers/api_provider.dart
@@ -41,7 +41,7 @@ class APIProvider {
         currencies.forEach((currencyCode, exchangeRate) {
           Rate rate = Rate(
             currencyCode: currencyCode,
-            exchangeRate: exchangeRate,
+            exchangeRate: exchangeRate.toDouble(),
             date: DateTime.parse(date),
           );
           ratesList.add(rate);
